Change schema to accept Float as input type for get line

+++ functions/src/apollo/schemas/baseSchema.ts
@@ -56,7 +56,7 @@ export const typeDefs = gql`
 
     input EndProductSpec {
         product: Product!
-        count: Int = 0
+        count: Float = 0
     }
 
     type ProductionStep {
